almost, asm algoritm generate blue fractal

+++ JuliaSetFractals/mainwindow.h
@@ -41,7 +41,7 @@ private:
     int height;
     void wheelEvent ( QWheelEvent * event );
     QImage GenerateJulia(std::complex<float> &c, int n_max =200, int width=400, int height=400);
-	QImage GenerateJuliaSSE(double cx, double cy, int n_max = 200, int width = 400, int height = 400);
+	QImage GenerateJuliaSSE(float cx, float cy, int n_max = 200, int width = 400, int height = 400);
 	QImage GenerateJuliaDoubles(double cx, double cy, int n_max = 200, int width = 400, int height = 400);
 	void drawMandelbrot(QImage *fractal,uint32_t *buffer, int xres, int yres);
 	__m128i IterateMandelbrot(__m128 a, __m128 b);
